moved contents of components into new sub-folder PAGES, edited administrators in api routes & models
