+++ client/src/App.js
@@ -2,11 +2,12 @@ import React, { Component } from "react";
 import { BrowserRouter as Router, Route, Link} from "react-router-dom";
 
 import "bootstrap/dist/css/bootstrap.min.css";
+import "./App.css";
 
-import Home from "./components/homeComponent";
-import About from "./components/aboutComponent";
-import Account from "./components/accountComponent";
-import Survey from "./components/surveyComponent";
+import Home from "./components/pages/home";
+import About from "./components/pages/about";
+import Login from "./components/pages/login";
+import Survey from "./components/pages/survey";
 
 
 class App extends Component {
@@ -22,7 +23,7 @@ class App extends Component {
                   <Link to="/about" className="nav-link">About </Link>
                   </li>
                   <li className="nav-item active">
-                  <Link to="/account" className="nav-link">Create Account </Link>
+                  <Link to="/login" className="nav-link">Login </Link>
                   </li>
                   <li className="nav-item active">
                   <Link to="/survey" className="nav-link">Survey</Link>
@@ -34,7 +35,7 @@ class App extends Component {
 
         <Route path="/" exact component={Home}/>
         <Route path="/about" exact component={About}/>
-        <Route path="/account" exact component={Account}/>
+        <Route path="/login" exact component={Login}/>
         <Route path="/survey" exact component={Survey}/>
       </Router>
